Add path conversion for MSYS2 environments
In a MSYS2-based environment on Windows (this includes the shell
provided by Git for Windows) Vim operates as if running on a Unix
system. The external tools used by vim-go on the other hand expect path
names in Windows path notation as command line arguments. When called
from a MSYS2 environment, they were supplied with a path in Unix notation
instead (e.g. /c/gopath/src/foo.go). To resolve this discrepancy, all
command line arguments passed to external tools must be translated into
Windows path notation in such an environment. Similarly, output created
by external tools must be converted back to Unix notation when it
contains path names.

This commit introduces functions for translating paths between internal
and external (native to the tool) notation. Path translation is
performed with the help of the cygpath utility provided by the Cygwin
environment on which MSYS2 is based. As Windows path names contain
backslashes, further escaping is often required when using the
conversion functions introduced here.

The path conversion functions are then used in the vim-go commands where
command line arguments for external tools are constructed or tool output
is parsed. The translation is a no-op on other platforms and when using
a native Vim build on Windows.